Release prep: mechanism hardening, trust/ops knobs, verification gates
- Fail-closed mechanism registry: explicit excluded flag (MECHANISM_INVALID),
  narrowing checks (FUNCTION_FAILED), proto-excluded field + redaction tests
- u64 session/object handle guards (32-bit-safe narrowing) incl. i686 leg
- SecretBytes PIN holders across FFI/auth/session paths; LOG_FORMAT knob
  (JSON default, opt-in plain) with startup-logging live leg
- CI/quality-gate fixes, tier runner header, OASIS inventory accuracy,
  CHANGELOG 0.2.0 entry
- Full workspace suite green; fmt/clippy clean; release dry-run passed

+++ crates/backend/src/ffi/exact_output_contract_tests.rs
@@ -125,7 +125,7 @@ fn authenticated_null_backing_accepts_aligned_and_partial_fixed_prefixes_without
             }),
         ] {
             let native = super::message_ops::build_message_init_mechanism(
-                cryptoki_sys::CKM_AES_GCM,
+                u64::from(cryptoki_sys::CKM_AES_GCM),
                 &parameter,
             )
             .unwrap();
